[Feat] 자주 가는 장소 수정 후 재조회 구현
