+++ TAMINGO/Features/MyPage/FavoritePlace/Service/FavoritePlacesService.swift
@@ -12,9 +12,10 @@ import Moya
 protocol FavoritePlacesServiceProtocol {
     func fetchPlaces() async throws -> [FavoritePlace]
     func createPlace(_ dto: PlaceRequestDTO) async throws -> Int
-    func updatePlace(placeId: Int, dto: PlaceRequestDTO) async throws -> Int
+    func updatePlace(placeId: Int, dto: PlaceRequestDTO) async throws
     func deletePlace(placeId: Int) async throws
 }
+
 
 final class FavoritePlacesService: FavoritePlacesServiceProtocol {
 
@@ -77,23 +78,17 @@ final class FavoritePlacesService: FavoritePlacesServiceProtocol {
 
 
     // MARK: - 수정
-    func updatePlace(placeId: Int, dto: PlaceRequestDTO) async throws -> Int {
+    func updatePlace(placeId: Int, dto: PlaceRequestDTO) async throws {
         let response = try await provider.requestAsync(.updatePlace(placeId: placeId, dto: dto))
 
-        let decoded = try JSONDecoder().decode(
-            BaseResponse<Int>.self,
-            from: response.data
-        )
-        
-        guard let result = decoded.result else {
+        guard (200..<300).contains(response.statusCode) else {
             throw APIError.server(
-                status: -1,
-                message: "서버 응답에 result가 없습니다."
+                status: response.statusCode,
+                message: "장소 수정에 실패했습니다."
             )
         }
-
-        return result
     }
+
 
 
     // MARK: - 삭제
