--- TAMINGO/Features/MyPage/Personalization/Model/ErrorLogSetting.swift
@@ -1,12 +1,0 @@
-//
-//  ErrorLog.swift
-//  TAMINGO
-//
-//  Created by 권예원 on 1/28/26.
-//
-
-import Foundation
-
-struct ErrorLogSetting {
-    let isEnabled: Bool
-}
